Extract PackageManager sub classes
